+++ src/tween.ts
@@ -1,58 +1,22 @@
-export const easings = {
-	linear: (x) => {
-		return x
-	},
-	easeInSine: (x) => {
-		return 1 - Math.cos((x * Math.PI) / 2)
-	},
-	easeOutSine: (x) => {
-		return Math.sin((x * Math.PI) / 2)
-	},
-	easeInOutSine: (x) => {
-		return -(Math.cos(Math.PI * x) - 1) / 2
-	},
-	easeInQuad: (x) => {
-		return x * x
-	},
-	easeOutQuad: (x) => {
-		return 1 - (1 - x) * (1 - x)
-	},
-	easeInOutQuad: (x) => {
-		return x < 0.5 ? 2 * x * x : 1 - Math.pow(-2 * x + 2, 2) / 2
-	},
-	easeInCubic: (x) => {
-		return x * x * x
-	},
-	easeOutCubic: (x) => {
-		return 1 - Math.pow(1 - x, 3)
-	},
-	easeInOutCubic: (x) => {
-		return x < 0.5 ? 4 * x * x * x : 1 - Math.pow(-2 * x + 2, 3) / 2
-	},
-	easeInQuart: (x) => {
-		return x * x * x * x
-	},
-	easeOutQuart: (x) => {
-		return 1 - Math.pow(1 - x, 4)
-	},
-	easeInOutQuart: (x) => {
-		return x < 0.5 ? 8 * x * x * x * x : 1 - Math.pow(-2 * x + 2, 4) / 2
-	},
-	easeInQuint: (x) => {
-		return x * x * x * x * x
-	},
-	easeOutQuint: (x) => {
-		return 1 - Math.pow(1 - x, 5)
-	},
-	easeInOutQuint: (x) => {
-		return x < 0.5 ? 16 * x * x * x * x * x : 1 - Math.pow(-2 * x + 2, 5) / 2
-	},
-	easeInExpo: (x) => {
-		return x === 0 ? 0 : Math.pow(2, 10 * x - 10)
-	},
-	easeOutExpo: (x) => {
-		return x === 1 ? 1 : 1 - Math.pow(2, -10 * x)
-	},
+const easings = {
+	linear: (x) => x,
+	easeInSine: (x) => 1 - Math.cos((x * Math.PI) / 2),
+	easeOutSine: (x) => Math.sin((x * Math.PI) / 2),
+	easeInOutSine: (x) => -(Math.cos(Math.PI * x) - 1) / 2,
+	easeInQuad: (x) => x * x,
+	easeOutQuad: (x) => 1 - (1 - x) * (1 - x),
+	easeInOutQuad: (x) => x < 0.5 ? 2 * x * x : 1 - Math.pow(-2 * x + 2, 2) / 2,
+	easeInCubic: (x) => x * x * x,
+	easeOutCubic: (x) => 1 - Math.pow(1 - x, 3),
+	easeInOutCubic: (x) => x < 0.5 ? 4 * x * x * x : 1 - Math.pow(-2 * x + 2, 3) / 2,
+	easeInQuart: (x) => x * x * x * x,
+	easeOutQuart: (x) => 1 - Math.pow(1 - x, 4),
+	easeInOutQuart: (x) => x < 0.5 ? 8 * x * x * x * x : 1 - Math.pow(-2 * x + 2, 4) / 2,
+	easeInQuint: (x) => x * x * x * x * x,
+	easeOutQuint: (x) => 1 - Math.pow(1 - x, 5),
+	easeInOutQuint: (x) => x < 0.5 ? 16 * x * x * x * x * x : 1 - Math.pow(-2 * x + 2, 5) / 2,
+	easeInExpo: (x) => x === 0 ? 0 : Math.pow(2, 10 * x - 10),
+	easeOutExpo: (x) => x === 1 ? 1 : 1 - Math.pow(2, -10 * x),
 	easeInOutExpo: (x) => {
 		return x === 0
 			? 0
@@ -61,12 +25,8 @@ export const easings = {
 				: x < 0.5 ? Math.pow(2, 20 * x - 10) / 2
 					: (2 - Math.pow(2, -20 * x + 10)) / 2
 	},
-	easeInCirc: (x) => {
-		return 1 - Math.sqrt(1 - Math.pow(x, 2))
-	},
-	easeOutCirc: (x) => {
-		return Math.sqrt(1 - Math.pow(x - 1, 2))
-	},
+	easeInCirc: (x) => 1 - Math.sqrt(1 - Math.pow(x, 2)),
+	easeOutCirc: (x) => Math.sqrt(1 - Math.pow(x - 1, 2)),
 	easeInOutCirc: (x) => {
 		return x < 0.5
 			? (1 - Math.sqrt(1 - Math.pow(2 * x, 2))) / 2
@@ -89,9 +49,7 @@ export const easings = {
 			? (Math.pow(2 * x, 2) * ((c2 + 1) * 2 * x - c2)) / 2
 			: (Math.pow(2 * x - 2, 2) * ((c2 + 1) * (x * 2 - 2) + c2) + 2) / 2
 	},
-	easeInBounce: (x) => {
-		return 1 - easings.easeOutBounce(1 - x)
-	},
+	easeInBounce: (x) => 1 - easings.easeOutBounce(1 - x),
 	easeOutBounce: (x) => {
 		const n1 = 7.5625
 		const d1 = 2.75
@@ -111,3 +69,5 @@ export const easings = {
 			: (1 + easings.easeOutBounce(2 * x - 1)) / 2
 	},
 }
+
+export default easings